AA-463: Extracted Related Functions Into their own file*: Pull some cohesive part of a function into its own file to reduce the size of a design element that is too big.
